Split geometric types to their own definition file
