+++ types/geometric.d.ts
@@ -1,0 +1,11 @@
+export interface ChartArea {
+  top: number;
+  left: number;
+  right: number;
+  bottom: number;
+}
+
+export interface Point {
+  x: number;
+  y: number;
+}
